+++ src/main/java/dev/jaxydog/astral/content/effect/AstralStatusEffects.java
@@ -1,3 +1,17 @@
+/*
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ *
+ * Copyright © 2024 Jaxydog
+ *
+ * This file is part of Astral.
+ *
+ * Astral is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ * Astral is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License along with Astral. If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package dev.jaxydog.astral.content.effect;
 
 import dev.jaxydog.astral.register.ContentRegistrar;
@@ -12,12 +26,29 @@ import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.passive.PufferfishEntity;
 import net.minecraft.util.math.random.Random;
 
-public final class CustomStatusEffects extends ContentRegistrar {
+/**
+ * Contains definitions of all modded-in status effects.
+ *
+ * @author Jaxydog
+ * @since 2.0.0
+ */
+public final class AstralStatusEffects extends ContentRegistrar {
 
-    public static final CustomStatusEffect SINISTER = new CustomStatusEffect("sinister",
+    /**
+     * The sinister effect.
+     *
+     * @since 2.0.0
+     */
+    public static final AstralStatusEffect SINISTER = new AstralStatusEffect("sinister",
         StatusEffectCategory.HARMFUL,
         0xE43727
     ) {
+
+        /**
+         * A scheduler instance used to delay effects for a tick.
+         *
+         * @since 2.0.0
+         */
         private final Scheduler scheduler = new Scheduler();
 
         @SuppressWarnings("RedundantCast")
@@ -77,6 +108,7 @@ public final class CustomStatusEffects extends ContentRegistrar {
                 }
             }
 
+            // Continue re-applying until the amplifier is zero.
             if (amplifier > 0) {
                 this.scheduler.queue(server -> entity.addStatusEffect(new StatusEffectInstance(SINISTER,
                     Math.min(20 * (amplifier), 200),