+++ docs/CHANGELOG.md
@@ -51,6 +51,7 @@ Essentially a rewrite of v1.*, adding/updating heaps of documentation and refact
     - Removed deprecated powers.
 - Updated data structure.
     - Renamed some fields and methods within the `MoonPhase` data type.
+- Updated effect content structure.
 
 ### Content changes